Manual Timetable Generation

+++ database/migrations/2024_03_25_133657_create_degree_module_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
     {
         Schema::create('degree_module', function (Blueprint $table) {
             $table->string('degree_id');
-            $table->integer('module_id');
+            $table->unsignedBigInteger('module_id');
             $table->timestamps();
 
             //Create composite key
